+++ NEWS.md
@@ -1,7 +1,7 @@
 # pathfindR (development version)
 
 ## Major Changes
-- Implemented greedy search in R
+- Implemented greedy active subnetwork search in R (breadth-first search)
 
 ## Minor changes and bug fixes