+++ NEWS.md
@@ -1,5 +1,7 @@
 # pathfindR (development version)
 
+- In `run_pathfindR()`, the default `iterations` was set back to 10 (the default for all other v1.x)
+
 ***
 
 # pathfindR 1.6.0