+++ NEWS.md
@@ -1,4 +1,4 @@
-# pathfindR (development version)
+# pathfindR 1.6.1
 
 - In `run_pathfindR()`, the default `iterations` was set back to 10 (the default for all other v1.x)